Place the interface request in moduleInit and execute it once every 10 seconds
